requirements-pigar-i.txt
# Automatically generated by https://github.com/damnever/pigar.
# pigar -i venv (ignores venv)

category_encoders == 2.5.0
dtreeviz == 1.3.7
fast_ml == 3.68
feature_engine == 1.4.1
featuretools == 1.14.0
img2pdf == 0.4.4
ipython == 8.2.0
matplotlib == 3.5.1
mlxtend == 0.21.0
nltk == 3.7
numpy == 1.22.3
pandas == 1.5.0
pyaml_env == 1.1.5
scikit_learn == 1.0.2
scipy == 1.8.0
seaborn == 0.11.2
streamlit == 1.12.2
umap_learn == 0.5.3
woodwork == 0.18.0
xgboost == 1.6.2
